Split DateTimeOffsetAssertionSpecs
